Quadtree bitfield operations made using type int64. Type int64 typedef'ed based on autoconf detected int type (long or long long). If long is not 64bits int64 will be really 32 bits and INT64_IS_REALLY32 will be defined.
git-svn-id: http://svn.osgeo.org/geos/trunk@222 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ source/headers/indexQuadtree.h
@@ -42,7 +42,7 @@ public:
 	static double maximumCommonMantissa(double d1, double d2);
 	DoubleBits(double nx);
 	double getDouble();
-	int biasedExponent();
+	int64 biasedExponent();
 	int getExponent();
 	void zeroLowerBits(int nBits);
 	int getBit(int i);
@@ -51,7 +51,8 @@ public:
 private:
 	double x;
 //	long long xBits;
-	long xBits;
+//	long xBits;
+	int64 xBits;
 };
 
 /**
